FEAT: combined OpenRouter/LM Arena report script, cleaner cost matching
Extract the LM Arena matching and cost-estimate logic out of
chat_with_model_choice.py into lmarena_scoring.py (needed since that app
file launches a Gradio server at import time, so nothing else could import
from it), and add build_model_report.py to fetch fresh OpenRouter pricing
and an LM Arena snapshot and combine them into one CSV report.

Also fixes false-positive score matches caused by thefuzz fragmenting
decimal version numbers (e.g. gpt-3.5 matching gpt-5.5 on a shared lone
digit), adds an organization-aware pre-filter to the matcher, and rounds
all numeric columns to 2 decimals so exported CSVs open cleanly in Excel.
